requirements.txt
streamlit
wordcloud
matplotlib
pillow
numpy
requests
beautifulsoup4
